# [IMP] CellPlugin: Avoid useless UPDATE_CELL