--- a/src/plugins/core/cell.ts
+++ b/src/plugins/core/cell.ts
@@ -162,7 +162,7 @@ export class CellPlugin extends CorePlugin<CoreState> implements CoreState {
       for (let col = zone.left; col <= zone.right; col++) {
         for (let row = zone.top; row <= zone.bottom; row++) {
           const cell = this.getters.getCell({ sheetId, col, row });
-          if (cell) {
+          if (cell?.isFormula || cell?.content) {
             this.dispatch("UPDATE_CELL", {
               sheetId: sheetId,
               content: "",
@@ -200,7 +200,6 @@ export class CellPlugin extends CorePlugin<CoreState> implements CoreState {
     for (let zone of recomputeZones(zones)) {
       for (let col = zone.left; col <= zone.right; col++) {
         for (let row = zone.top; row <= zone.bottom; row++) {
-          // commandHelpers.updateCell(sheetId, col, row, { style: undefined});
           this.dispatch("UPDATE_CELL", {
             sheetId,
             col,