@robodoo 2024-11-28
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/5283.png)](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/5283)